requirements.txt
Flask>=3.0
torch>=2.2
torchvision>=0.17
torchaudio>=2.2
Pillow>=10.0
numpy>=1.26
